chore(menu) : 테스트 코드 이동
